Update according to MSFT comments
1. Don't need to migrate for single buffer pool mode
2. Move buffer setting migration to another file
3. Remove unnecessary code/upgrading flows
